Move parser up a directory.
