+++ npm-packages/meteor-vite/src/meteor/package/components/MeteorPackage.ts
@@ -1,7 +1,7 @@
 import Path from 'path';
 import { MeteorViteError } from '../../../vite/error/MeteorViteError';
-import type { ModuleList, ParsedPackage } from '../parser/Parser';
-import { parseMeteorPackage } from '../parser/Parser';
+import type { ModuleList, ParsedPackage } from '../../parser/Parser';
+import { parseMeteorPackage } from '../../parser/Parser';
 import { SerializationStore } from '../SerializationStore';
 import PackageExport from './PackageExport';
 import { PackageSubmodule } from './PackageSubmodule';
