+++ npm-packages/meteor-vite/src/meteor/parser/Parser.ts
@@ -9,8 +9,8 @@ import {
 } from '@babel/types';
 import FS from 'fs/promises';
 import { inspect } from 'util';
-import Logger from '../../../Logger';
-import { MeteorViteError } from '../../../vite/error/MeteorViteError';
+import Logger from '../../Logger';
+import { MeteorViteError } from '../../vite/error/MeteorViteError';
 import {
     KnownModuleMethodNames,
     MeteorPackageProperty,
