+++ npm-packages/meteor-vite/test/__mocks/index.ts
@@ -1,6 +1,6 @@
 import FS from 'fs/promises';
 import Path from 'path';
-import { ModuleList, PackageScopeExports } from '../../src/meteor/package/parser/Parser';
+import { ModuleList, PackageScopeExports } from '../../src/meteor/parser/Parser';
 
 export const AllMockPackages: MockModule<ModuleList>[] = [];
